Fixes #13311: Moved unit tests inside the permission registry and added the test for api operations on base api tests
